Create test suuport for reusable fakes
