+++ projection.go
@@ -1,10 +1,6 @@
 package es
 
 // Projection abstraction
-type Projection struct {
-	version int64
-}
-
-func (v *Projection) setVersion(version int64) {
-	v.version = version
+type Projection interface {
+	Reduce(event Event)
 }
